+++ src/api/action/user/mod.rs
@@ -3,3 +3,4 @@ pub(crate) mod delete;
 pub(crate) mod list;
 pub(crate) mod me;
 pub(crate) mod update;
+pub(crate) mod validation;