+++ src/api/mod.rs
@@ -1,3 +1,4 @@
 pub(crate) mod action;
 pub(crate) mod dto;
 pub(crate) mod server;
+pub(crate) mod validation;